Initial commit, notes about development
This is a fork of the "Spartan" JSON Resume theme. I'm unlikely to
officially publish this since I'm just trying to meet my own needs.

NOTE: As of the time of this commit, the ONLY WAY to develop this theme
locally is to use version 1.2.8 of the resume CLI (npm i -g resume-cli@1.2.8

NOTE: As of the time of this commit, here are the steps that must be
taken to work on this:

    1. Specifically install resume CLI version 1.2.8 (npm i -g
resume-cli@1.2.8)
    2. Add your resume.json directly to the root folder of the theme
(see here: https://jsonresume.org/themes/)
    3. Don't forget to install dependencies like you just did for an
hour :^(
    4. Once dependencies are installed, `resume serve` from theme folder